Added new API to IEditor "SetLexerName" which allows plugin to set differnet syntax highlight
git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@3594 9da81c78-c036-0410-9e1f-a2b0375e4b5a

+++ LiteEditor/context_cpp.cpp
@@ -1998,7 +1998,7 @@ void ContextCpp::ApplySettings()
 
 	// Update the control
 	LEditor &rCtrl = GetCtrl();
-	rCtrl.SetLexer(lexPtr->GetLexerId());
+	rCtrl.SetLexer((int)lexPtr->GetLexerId());
 
 	wxString keyWords = lexPtr->GetKeyWords(0);
 	keyWords.Replace(wxT("\n"), wxT(" "));
